Revert "Refactor Delete Command Class."
This reverts commit 5800a27dac6d0f4ce7f7026389f75037de6f3ea6.
